services/langsmith-exporter/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langsmith>=0.1.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
